Doyok

+++ app/Http/Controllers/TAController.php
@@ -191,29 +191,27 @@ class TAController extends Controller
     {
         $tugas_akhir = TA::find($id);
         $data = $request->all();
+        $hapus = $tugas_akhir->praproposal;
         if ($request->file('praproposal')) {
-            $TA = TA::latest()->get();
-            $mhs_id = Mahasiswa::where('id', $request->mahasiswa)->get()->first();
-            // dd($mhs_id);
-            $nim = $mhs_id->nim;
+           $mhs_id = Mahasiswa::where('id', $tugas_akhir->mahasiswa_id)->get()->first();
+           $nim = $mhs_id->nim;
+        //    dd($request->tahunAkademik);
             $file = $request->file('praproposal');
             $filename = 'TA' . '_' . $nim . '_' . time() . '.' . $file->getClientOriginalExtension();
             $path = $request->file('praproposal')->storeAS('public/assets/file/PraproposalTA/', $filename);
             $data = [
-                'mahasiswa_id' => $request->mahasiswa,
                 'judulTA' => $request->judulTA,
                 'instansi' => $request->instansi,
                 'pembimbing1_id' => $request->pembimbing1_id,
                 'pembimbing2_id' => $request->pembimbing2_id,
                 'ket' => $request->ket,
-                'status_id' => $request->status,
-                'thnAkad_id' => $request->tahunAkademik,
+                'status_id' => $request->status_id,
                 'praproposal' => $filename,
             ];
+            File::delete(public_path('storage/assets/file/PraproposalTA/' . $hapus . ''));
         } else {
             $data['praproposal'] = $tugas_akhir->praproposal;
         }
-
         $tugas_akhir->update($data);
         Alert::success('Berhasil', 'Berhasil Mengubah Data TA');
         return redirect(route('TA.index'));
@@ -228,8 +226,14 @@ class TAController extends Controller
     public function destroy($id)
     {
         $tugas_akhir = TA::find($id);
-        $tugas_akhir->delete();
-        Alert::success('Berhasil', 'Berhasil hapus data Tugas Akhir');
+        if ($tugas_akhir != null) {
+            File::delete(public_path('storage/assets/file/PraproposalTA/' . $tugas_akhir->praproposal . ''));
+            // dd($tugas_akhir->praproposal);
+            $tugas_akhir->delete();
+            Alert::success('Berhasil', 'Berhasil hapus data Tugas Akhir');
+            return back();
+        }
+        Alert::warning('Gagal', 'Data Tugas Akhir Gagal Dihapus');
         return back();
     }
     public function download($filename)
